+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+- Do not expose `AuthorizationResource` anymore via REST API (always identify by resource class and identifier)
+- Do not explicitly register resources anymore. Create `AuthorizationResource` implicitly when adding a resource action grant
+for a resource that does not yet exist.
+
 ## v0.5.4
 
 - Add `removeResourceActionGrants` to PHP API